yolo-tune/requirements.txt
label-studio
tifffile
Pillow
ultralytics
ipykernel